+++ src/Expanders.jl
@@ -562,28 +562,14 @@ function Selectors.runner(::Type{ExampleBlocks}, x, page, doc)
     content = []
     input   = droplines(x.code)
 
-    # Special-case support for displaying SVG and PNG graphics. TODO: make this more general.
-    output = if showable(MIME"text/html"(), result)
-        Documents.RawHTML(Base.invokelatest(stringmime, MIME"text/html"(), result))
-    elseif showable(MIME"image/svg+xml"(), result)
-        Documents.RawHTML(Base.invokelatest(stringmime, MIME"image/svg+xml"(), result))
-    elseif showable(MIME"image/png"(), result)
-        Documents.RawHTML(string("<img src=\"data:image/png;base64,", Base.invokelatest(stringmime, MIME"image/png"(), result), "\" />"))
-    elseif showable(MIME"image/webp"(), result)
-        Documents.RawHTML(string("<img src=\"data:image/webp;base64,", Base.invokelatest(stringmime, MIME"image/webp"(), result), "\" />"))
-    elseif showable(MIME"image/gif"(), result)
-        Documents.RawHTML(string("<img src=\"data:image/gif;base64,", Base.invokelatest(stringmime, MIME"image/gif"(), result), "\" />"))
-    elseif showable(MIME"image/jpeg"(), result)
-        Documents.RawHTML(string("<img src=\"data:image/jpeg;base64,", Base.invokelatest(stringmime, MIME"image/jpeg"(), result), "\" />"))
-    else
-        Markdown.Code(Documenter.DocTests.result_to_string(buffer, result))
-    end
+    # Generate different  in different formats and let each writer select
+    output = Base.invokelatest(Utilities.display_dict, result)
 
     # Only add content when there's actually something to add.
     isempty(input)  || push!(content, Markdown.Code("julia", input))
-    isempty(output.code) || push!(content, output)
+    isempty(output) || push!(content, output)
     # ... and finally map the original code block to the newly generated ones.
-    page.mapping[x] = Markdown.MD(content)
+    page.mapping[x] = Documents.MultiOutput(content)
 end
 
 # @repl