+++ src/Utilities/Utilities.jl
@@ -635,6 +635,20 @@ function mdparse(s::AbstractString; mode=:single)
     end
 end
 
+# Capturing output in different representations similar to IJulia.jl
+import Base64: stringmime
+function display_dict(x)
+    out = Dict{MIME,Any}()
+    # Always generate text/plain
+    out[MIME"text/plain"()] = stringmime(MIME"text/plain"(), x)
+    for m in [MIME"text/html"(), MIME"image/svg+xml"(), MIME"image/png"(),
+              MIME"image/webp"(), MIME"image/gif"(), MIME"image/jpeg"(),
+              MIME"text/latex"()]
+        showable(m, x) && (out[m] = stringmime(m, x))
+    end
+    return out
+end
+
 include("DOM.jl")
 include("MDFlatten.jl")
 include("TextDiff.jl")